Add files via upload
A password generator is a useful tool that generates strong and random passwords for users. This project aims to create a password generator application using Python, allowing users to specify the length and complexity of the password.

function : 
   - Accepts two parameters: length (integer) and complexity (string).
   - Depending on the complexity level chosen (weak, medium, or strong), it selects appropriate character sets from the string module (ascii_letters, digits, 
      punctuation, ascii_uppercase, ascii_lowercase).
   - Generates a password of the specified length using random.choice to pick characters from the selected character set.

Main Function :
   - Displays a welcome message and prompts the user to enter the desired length of the password and its complexity.
   - Calls generate_password with user-provided inputs.
   - Prints the generated password.

How to use :
   - Run the script.
   - Enter the desired password length.
   - Choose the complexity level (weak, medium, or strong).
   - The script then generates and displays a password based on the user's inputs.

This script provides a basic implementation of a password generator application in Python, allowing users flexibility in determining the password length and complexity according to their security needs.